fix: add useResizeTextarea - see #6920
---

## Описание

Линтер компилятора ругается на мутации внутри Textarea, WriteBar

> Mutating a value returned from a function whose return value should not be mutated

## Изменение

Вынес общий хук useResizeTextarea
